03_speaker_database/requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
Flask-HTTPAuth==4.7.0
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-pubsub==2.13.4
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.6
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
